feat: add some code and refactoring
fix all issues detected and add code, refactoring and add logs

+++ src/main/java/com/suehay/fsastorageservice/service/ProductServiceImpl.java
@@ -4,12 +4,15 @@ import com.suehay.fsastorageservice.model.entity.Product;
 import com.suehay.fsastorageservice.model.request.GenericPageRequest;
 import com.suehay.fsastorageservice.model.response.GenericResponse;
 import com.suehay.fsastorageservice.repository.ProductRepository;
+import com.suehay.fsastorageservice.util.Logger;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -17,22 +20,30 @@ import java.util.List;
 public class ProductServiceImpl implements ProductService {
     private final FileStorageService fileStorageService;
     private final ProductRepository productRepository;
+    private final Logger logger;
+    private final String CONTEXT = "ProductService::";
 
     @Override
     public GenericResponse<List<Product>> getProducts(GenericPageRequest<String> request) {
+        var FUNCTION_CONTEXT = CONTEXT + "getProducts";
         var response = new GenericResponse<List<Product>>();
+
         if (request.getFilter() != null) {
+            logger.info(FUNCTION_CONTEXT, "Getting products...");
             response.setData(productRepository.findAllByCategoryNameContains(
                     PageRequest.of(request.getPage(), request.getSize()),
                     request.getFilter()).stream().toList());
         } else {
+            logger.info(FUNCTION_CONTEXT, "Getting products...");
             response.setData(productRepository.findAll(PageRequest.of(request.getPage(), request.getSize())).stream().toList());
         }
+
         if (response.getData().isEmpty()) {
-            response.setMessage("No products found with category name: " + request.getFilter());
-            response.setStatus("404");
-            response.setError("Not Found");
+            logger.info(FUNCTION_CONTEXT, "No products found.");
+            response.setMessage("No products found.");
+            response.setStatus("200");
         } else {
+            logger.info(FUNCTION_CONTEXT, "Products found.");
             response.setMessage("Products found with category name: " + request.getFilter());
             response.setStatus("200");
         }
@@ -40,14 +51,32 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public GenericResponse<Product> getProductById(String id) {
+    public GenericResponse<Product> getProduct(Long id, String title) {
+        var FUNCTION_CONTEXT = CONTEXT + "getProduct";
+
         var response = new GenericResponse<Product>();
-        var product = productRepository.findById(Long.parseLong(id));
+        Optional<Product> product;
+
+        if (Objects.isNull(id) && Objects.isNull(title)) {
+            logger.info(FUNCTION_CONTEXT, "No id or title provided. Error in request.");
+            return new GenericResponse<>("No id or title provided.", "You need to provide a nonnull id or title to get a product.", "400", null);
+        }
+
+        if (!Objects.isNull(id)) {
+            logger.info(FUNCTION_CONTEXT, "Id provided.");
+            product = productRepository.findById(id);
+        } else {
+            logger.info(FUNCTION_CONTEXT, "Title provided.");
+            product = productRepository.findByTitle(title);
+        }
+
         if (product.isPresent()) {
+            logger.info(FUNCTION_CONTEXT, "Founded product.");
             response.setData(product.get());
-            response.setMessage("Product found with id: " + id);
+            response.setMessage("Product found.");
             response.setStatus("200");
         } else {
+            logger.info(FUNCTION_CONTEXT, "Product not found.");
             response.setMessage("No product found with id: " + id);
             response.setStatus("404");
             response.setError("Not Found");
@@ -57,8 +86,12 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public GenericResponse<Product> saveProduct(Product product) {
+        var FUNCTION_CONTEXT = CONTEXT + "saveProduct";
         var response = new GenericResponse<Product>();
         var addresses = fileStorageService.getAddressByNames(product.getImages());
+
+        logger.info(FUNCTION_CONTEXT, "Saving product...");
+
         product.setImages(addresses);
         response.setData(productRepository.save(product));
         response.setMessage("Product saved with id: " + product.getId());
@@ -68,15 +101,21 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public GenericResponse<Product> updateProduct(Product product) {
+        var FUNCTION_CONTEXT = CONTEXT + "updateProduct";
         var response = new GenericResponse<Product>();
         var productData = productRepository.findById(product.getId());
+
+        logger.info(FUNCTION_CONTEXT, "Updating product...");
+
         if (productData.isPresent()) {
+            logger.info(FUNCTION_CONTEXT, "Product saved.");
             var addresses = fileStorageService.getAddressByNames(product.getImages());
             product.setImages(addresses);
             response.setData(productRepository.save(product));
             response.setMessage("Product updated with id: " + product.getId());
             response.setStatus("200");
         } else {
+            logger.info(FUNCTION_CONTEXT, "Product not found.");
             response.setMessage("No product found with id: " + product.getId());
             response.setStatus("404");
             response.setError("Not Found");
@@ -85,43 +124,29 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public GenericResponse<String> deleteProductByName(String title) {
+    public GenericResponse<String> deleteProduct(Long id, String title) {
+        var FUNCTION_CONTEXT = CONTEXT + "deleteProduct";
         var response = new GenericResponse<String>();
-        var repositoryResponse = productRepository.deleteByTitle(title);
-        if (repositoryResponse) {
-            response.setMessage("Product deleted with title: " + title);
-            response.setData(title);
-            response.setStatus("200");
-        } else {
-            response.setMessage("No product found with title: " + title);
-            response.setStatus("404");
-            response.setError("Not Found");
-        }
-        return response;
-    }
 
-    @Override
-    public GenericResponse<Product> getProductByName(String name) {
-        var response = new GenericResponse<Product>();
-        var product = productRepository.findByTitle(name);
-        if (product.isPresent()) {
-            response.setData(product.get());
-            response.setMessage("Product found with title: " + name);
-            response.setStatus("200");
-        } else {
-            response.setMessage("No product found with title: " + name);
-            response.setStatus("404");
-            response.setError("Not Found");
-        }
-        return response;
-    }
+        logger.info(FUNCTION_CONTEXT, "Deleting product...");
 
-    @Override
-    public GenericResponse<String> deleteProduct(String id) {
-        var response = new GenericResponse<String>();
-        productRepository.deleteById(Long.parseLong(id));
+        if (Objects.isNull(id) && Objects.isNull(title)) {
+            logger.info(FUNCTION_CONTEXT, "No id or title provided. Error in request.");
+            return new GenericResponse<>("No id or title provided.", "You need to provide a nonnull id or title to delete a product.", "400", null);
+        }
+
+        if (!Objects.isNull(id)) {
+            logger.info(FUNCTION_CONTEXT, "Id provided.");
+            productRepository.deleteById(id);
+        } else {
+            logger.info(FUNCTION_CONTEXT, "Title provided.");
+            productRepository.deleteByTitle(title);
+        }
+
+        logger.info(FUNCTION_CONTEXT, "Product deleted.");
+
         response.setMessage("Product deleted with id: " + id);
-        response.setData(id);
+        response.setData(String.valueOf(id));
         response.setStatus("200");
         return response;
     }
